+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
-## [1.5.19](https://github.com/globocom/GloboDNS/releases/tag/1.5.19) (**/02/17)
-##### - Bug fixies
+## [1.5.22](https://github.com/globocom/GloboDNS/releases/tag/1.5.22)
+#### - Bug fixies
+ * Record search field had the same name as domain search field
+#### - Enhancements
+ * Supporting the use of views
+ * Task to migrate viewless zones to a default view (when using view, all zones must be in views)
+ * Adding filed 'schedule_date' to 'Schedule Export' JSON output
+
+## [1.5.19](https://github.com/globocom/GloboDNS/releases/tag/1.5.19) (07/02/17)
+##### - Bug fixies 
  * Validation of record TXT was being applied to all record types
  * SRV record must have 'Priority', 'Weight' and 'Port' fields
-##### - Enhancements
+##### - Enhancements 
  * Changed subject exporter/importer mailer, so that the emails received of different environments won't be agrouped
 
 ## [1.5.14](https://github.com/globocom/GloboDNS/releases/tag/1.5.14) (19/01/17)